requirements.txt
fastapi
pydantic
sqlalchemy
axios
sqladmin
pyjwt
pwdlib[argon2]
jwt
PyJWT 
python-jose